add tooltip for award
